CoinNeko_Django/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
packaging==26.0
pillow==12.1.1
PyJWT==2.12.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
